Utility function to identify records for deletion

Filters for records matching a given string.

Usage

remove_deletions(x, val = "Delete")

Arguments

x

input vector

val

The value to check for inequality. Defaults to 'Delete'

Value

logical vector

Details

To be used within dplyr::filter(). The function returns a logical vector with TRUE resulting from values that are not equal to the val argument. Also protects from NA values.

Used within verbs such as tidyselect::all_of() this can work effectively across all columns in a data frame. See examples

Examples

if (FALSE) { # \dontrun{
data |> filter(if_all(everything(), remove_deletions))
} # }